# public/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: GhostClass API
  version: 4.5.0
  description: |
    **GhostClass API** provides endpoints for authentication, profile synchronization,
    attendance integrations with EzyGo, telemetry, and build provenance.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (CRON_SECRET)**: For automated cron jobs
       - Add `Authorization: Bearer <CRON_SECRET>` header
       - Used by automated sync processes

    2. **Session Cookie (SupabaseAuth)**: For authenticated users
       - Automatically included when logged in through the web interface
       - Managed by Supabase Auth

    ## Rate Limiting

    All endpoints are rate-limited to prevent abuse:

    - `/api/auth/save-token` & `/api/cron/sync`: Configurable via environment variables
    - `/api/backend/*`: Configurable via `PROXY_RATE_LIMIT_REQUESTS` and `PROXY_RATE_LIMIT_WINDOW`
    - Contact Form (Server Action): Configurable via environment variables

    Rate limit information is returned in response headers:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Base URLs

    - Production: Configured via `NEXT_PUBLIC_APP_URL`
    - Development: `http://localhost:3000`

    ## Error Handling

    All error responses follow a consistent format with appropriate HTTP status codes and descriptive messages.

  contact:
...

...
servers:
...
paths:
...
  /api/user/accept-terms:
    post:
      tags:
        - Mobile
      summary: Accept terms for mobile user
      description: |
        Updates `terms_version` and `terms_accepted_at` for authenticated mobile users.
      operationId: acceptTermsMobile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - version
              properties:
                version:
                  type: string
      responses:
        "200":
          description: Terms acceptance stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  version:
                    type: string
        "400":
          description: Missing version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to update terms acceptance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong
        message:
          type: string
          description: Detailed error message
      example:
        error: "Internal Server Error"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token authentication using CRON_SECRET environment variable.
        Used for automated cron jobs.

        Example: `Authorization: Bearer <CRON_SECRET>`

    SupabaseAuth:
      type: apiKey
      in: cookie
      name: sb-*
      description: |
        Session cookie authentication managed by Supabase Auth.
        Automatically included when users are logged in through the web interface.
        Cookie names typically start with 'sb-' prefix.
...